% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analytics}
\alias{analytics}
\title{Analytics endpoints from Homebrews JSON API}
\format{A data frame with 19 rows and 1 variable:
\describe{
  \item{analytics}{a list of all analytics endpoints}
}}
\source{
\url{https://formulae.brew.sh/docs/api/}
}
\usage{
analytics
}
\description{
A dataset containing a list of 19 analytics endpoints. This doesn't include
formula endpoints e.g. /formula/sqlite.
}
\keyword{datasets}
