% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_response.R
\name{print_response}
\alias{print_response}
\title{Print the result of a GitHub API call}
\usage{
print_response(x, ...)
}
\arguments{
\item{x}{The result object.}

\item{...}{Ignored.}
}
\value{
The JSON result.
}
\description{
Print the result of a GitHub API call
}
