% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_to_df.R
\name{recipe_to_df}
\alias{recipe_to_df}
\title{Convert a recipe in beerXML format to an R data frame.}
\usage{
recipe_to_df(beerXML_file)
}
\arguments{
\item{beerXML_file}{A recipe in beerXML format.}
}
\value{
A data frame.
}
\description{
Convert a recipe in beerXML format to an R data frame.
}
